requirements.txt
# Literature Agent System Requirements

# 核心依赖
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0

# HTTP请求
requests>=2.31.0
urllib3>=2.0.0

# PDF处理和OCR（第二阶段）
pdf2image>=1.16.0
Pillow>=10.0.0
PyMuPDF>=1.23.0  # fitz，用于PDF图片提取

# 可选依赖（推荐安装）
tqdm>=4.65.0  # 进度条显示

# 开发依赖（可选）
# pytest>=7.4.0  # 测试框架
# black>=23.0.0  # 代码格式化
# flake8>=6.0.0  # 代码检查

# 注意：
# - pdf2image 需要系统安装 poppler-utils
#   macOS: brew install poppler
#   Ubuntu: sudo apt-get install poppler-utils
#   Windows: 下载 poppler 并添加到PATH
